% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{closeness.inv}
\alias{closeness.inv}
\title{Closeness}
\usage{
closeness.inv(X, Y)
}
\value{
Return
}
\description{
A
}
\details{
One of the methods to compare two networks on their node-specific differences.
}
\examples{
closeness.inv(X,Y)
}
