% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{comp.plot}
\alias{comp.plot}
\title{Provides plots allowing for visual comparison of two networks}
\usage{
comp.plot(
  A,
  B,
  methodlist,
  thresh = 0.05,
  networkA = TRUE,
  networkB = TRUE,
  networkAtitle = "Network A",
  networkBtitle = "Network B",
  cluster = TRUE,
  negcol = "red",
  poscol = "blue",
  multiplier = 4,
  curved = TRUE,
  layout = igraph::layout.auto,
  vSize = 16,
  tSize = 0.8
)
}
\arguments{
\item{A, B}{input data tables from which the adjacency matrices will be generated, to be provided in form of matrices, arrays, data frames or tibbles}

\item{methodlist}{a list specifying the method which is used to estimate and create the adjacency matrix; see details for further information}

\item{thresh}{a number between 0 and 1 (default is set to 0.05) specifying the singificance level: if the p-value corresponding to an edge weight is greater than \code{thresh}, the corresponding edge weight is not considered to be significant and thus set to zero}

\item{networkA}{Boolean, indicating whether the network corresponding to the first data set shall be plotted}

\item{networkB}{Boolean, indicating whether the network corresponding to the second data set shall be plotted}

\item{networkAtitle}{a character, indicating the title of the first network}

\item{networkBtitle}{a character, indicating the title of the second network}

\item{cluster}{a Boolean, indicating whether the networks shall be plotted along with the derived communities/clusters}

\item{negcol}{a character, specifying the color which is used to represent edges with a negative correlation}

\item{poscol}{a character, specifying the color which is used to represent edges with a positive correlation}

\item{multiplier}{a number, representing the factor with which an edge weight is multiplied in order to regulate the thickness of the drawn edges}

\item{curved}{a Boolean, indicating whether the edges should be drawn using a curved or solid line}

\item{layout}{a function specifying the layout of the networks; see details for possible options and further information}

\item{vSize}{a number, specifying the size of the nodes}

\item{tSize}{a number, specifying the node label size}
}
\value{
a plot of the two specified networks, which allows for visual comparison
}
\description{
This function provides plots allowing for visual comparison of two networks, which are given by adjacency matrices that are produced out of input data tables, using an estimation method specified by the user.
}
\details{
This function provides plots allowing for visual comparison of two networks, which are given by adjacency matrices that are produced out of input data tables, using an estimation method specified by the user. The network estimation method has to be specified in form of a list in the \code{methodlist} argument. Currently, the following estimation methods are supported:
\itemize{
\item{\code{list("Spearman")} \cr Edge weights are estimated using Spearman correlation, where unadjusted p-values are employed to determine significance. To apply this method, the expression \code{list("Spearman")} has to be provided in the \code{methodlist} argument.}
\item{\code{list("Spearman.adj",adjustment method)} \cr Edge weights are estimated using Spearman correlation, where p-values adjusted for multiple testing are employed to determine significance. To apply this method, the expression \code{list("Spearman.adj",adjustment method)} has to be provided in the \code{methodlist} argument, where \code{adjustment method} has to be one of the options for multple testing adjustment provided by the standard \code{p.adjust} R function, i.e. one of \code{"BH"}, \code{"bonferroni"}, \code{"BY"}, \code{"fdr"}, \code{"hochberg"}, \code{"holm"} or \code{"hommel"}.}
\item{\code{list("PCSpearman")} \cr Edge weights are estimated using partial Spearman correlation, where unadjusted p-values are employed to determine significance. To apply this method, the expression \code{list("PCSpearman")} has to be provided in the \code{methodlist} argument.}
\item{\code{list("PCSpearman.adj",adjustment method)} \cr Edge weights are estimated using partial Spearman correlation, where p-values adjusted for multiple testing are employed to determine significance. To apply this method, the expression \code{list("PCSpearman.adj",adjustment method)} has to be provided in the \code{methodlist} argument, where \code{adjustment method} has to be one of the options for multple testing adjustment provided by the standard \code{p.adjust} R function, i.e. one of \code{"BH"}, \code{"bonferroni"}, \code{"BY"}, \code{"fdr"}, \code{"hochberg"}, \code{"holm"} or \code{"hommel"}.}
\item{\code{list("DistCorr")} \cr Edge weights are estimated using distance correlation, where unadjusted p-values are employed to determine significance. To apply this method, the expression \code{list("DistCorr")} has to be provided in the \code{methodlist} argument. Note that the calculations may require larger computation times, as a permutation test is involved to derive the corresponding p-values for the distance correlations.}
\item{\code{list("DistCorr.adj",adjustment method)} \cr Edge weights are estimated using distance correlation, where p-values adjusted for multiple testing are employed to determine significance. To apply this method, the expression \code{list("DistCorr.adj",adjustment method)} has to be provided in the \code{methodlist} argument, where \code{adjustment method} has to be one of the options for multple testing adjustment provided by the standard \code{p.adjust} R function, i.e. one of \code{"BH"}, \code{"bonferroni"}, \code{"BY"}, \code{"fdr"}, \code{"hochberg"}, \code{"holm"} or \code{"hommel"}. Note that the calculations may require larger computation times, as a permutation test is involved to derive the corresponding p-values for the distance correlations.}
\item{\code{list("EBICglasso",correlation type,tuning parameter)} \cr Edge weights are estimated using the EBICglasso approach. To apply this method, the expression \code{list("EBICglasso",correlation type,tuning parameter)} has to be provided in the \code{methodlist} argument. Here, \code{correlation type} has to be one of the association options provided by the standard \code{cor} R function, i.e. one of \code{"kendall"}, \code{"pearson"} or \code{"spearman"}. Moreover, \code{tuning parameter} has to be a number specifying the EBIC tuning parameter \eqn{\gamma}. Typical choices include values between 0 and 0.5, where smaller values usually lead to a higher sensitivity in that more edges are included into the network. \cr Note that for EBICglasso, an additional specification of the \code{thresh} argument is obsolete, as it is not used for the application of the method.}
}
The following options (functions) from the \code{igraph} R package are provided to specify the layout of the plots in the \code{layout} argument:
\itemize{
\item{\code{igraph::layout.auto} (default)}
\item{\code{igraph::layout.circle}}
\item{\code{igraph::layout.davidson.harel}}
\item{\code{igraph::layout.drl}}
\item{\code{igraph::layout.fruchterman.reingold}}
\item{\code{igraph::layout.gem}}
\item{\code{igraph::layout.graphopt}}
\item{\code{igraph::layout.grid}}
\item{\code{igraph::layout.kamada.kawai}}
\item{\code{igraph::layout.lgl}}
\item{\code{igraph::layout.mds}}
\item{\code{igraph::layout.reingold.tilfort}}
\item{\code{igraph::layout.star}}
\item{\code{igraph::layout.svd}}
}
}
\examples{
comp.plot(ExDataA,ExDataB,methodlist=list("Spearman"))
comp.plot(ExDataA,ExDataB,methodlist=list("PCSpearman.adj","bonferroni"),
layout=igraph::layout.circle,curved=FALSE)
comp.plot(ExDataA,ExDataB,methodlist=list("EBICglasso","spearman",0.1),
layout=igraph::layout.fruchterman.reingold,curved=FALSE)
comp.plot(ExDataA,ExDataB,methodlist=list("EBICglasso","pearson",0.05),
layout=igraph::layout.star,cluster=FALSE)
}
