% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{create.adjacency.matrix}
\alias{create.adjacency.matrix}
\title{Create Adjacency Matrix}
\usage{
create.adjacency.matrix(A, methodlist, thresh = 0.05)
}
\arguments{
\item{A}{a matrix.}

\item{methodlist}{the methods used to create the adjacency matrix.}

\item{thresh}{the threshold under which the values of the correlations are relevant.}
}
\value{
The adjacency matrix resulting from the matrix A.
}
\description{
The method creates an adjacency matrix out of the matrix A with the methods specified in the attribute methodlist.
}
\details{
To show the meaning of a certain value in the adjacency matrix, the titles of the columns of the matrix A are used
as titles of the rows and the columns of the adjacency matrix.
}
\examples{
create.adjacency.matrix(A,list("Spearman"))
create.adjacency.matrix(A,list("DistCorr.adj", "holm"))
}
