% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R, R/app.R
\name{create.app}
\alias{create.app}
\title{Create App}
\usage{
create.app()

create.app()
}
\description{
Creates the Shiny App of the Differential Network Tests

Creates the Shiny App of the Differential Network Tests
}
