% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{create.graph}
\alias{create.graph}
\title{Create Graph}
\usage{
create.graph(A, methodlist, thresh = 0.05)
}
\arguments{
\item{A}{a matrix.}

\item{methodlist}{the methods used to create the adjacency matrix.}

\item{thresh}{the threshold under which the values of the correlations are relevant.}
}
\value{
create.graph returns a list of the adjacency matrix, the graph, the graph communities, the graph clustering and
the graphs vertex degrees, number of edges, number of clusters and number of isolated nodes. Followed by the minimum
spanning tree, its communities, its clustering and the MSTs vertex degrees, number of edges, number of clusters and
number of isolated nodes.
}
\description{
The method creates an graph out of the matrix A.
}
\details{
To create the graph (adjacency) out of any matrix A this matrix needs to be converted into an adjacency matrix. This is
done by the method create.adjacency.matrix. Then many of the graphs characteristics are evaluated and saved in variables.
With the help of Prim's Algorithm the graph (adjacency) is converted into a minimum spanning tree (short: MST) of which
the same characteristics are evaluated.
}
\examples{
create.graph(A,list("Spearman"))
}
