% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{degree.inv}
\alias{degree.inv}
\title{Degree differneces}
\usage{
degree.inv(X, Y)
}
\arguments{
\item{X, Y}{are matrices}
}
\value{
A vector of the differneces of the degrees of the same nodes in two different networks.
}
\description{
The method calculates the differences between the degrees of the nodes of two matrices.
}
\details{
One of the methods to compare two networks on their node-specific differences.
}
\examples{
degree.inv(X,Y)

}
