% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{diff.num}
\alias{diff.num}
\title{Differneces between two graphs}
\usage{
\method{diff}{num}(A, B)
}
\arguments{
\item{A, B}{are results from the function create.graph.}
}
\value{
A list of the differences in the number of edges, clusters and isolated nodes of the graph and the corresponding
minimum spanning tree.
}
\description{
The method calculates differences between two graphs.
}
\details{
One of the methods to compare two networks on their overall network characteristics.
}
\examples{
diff.num(create.graph(A, list("Spearman")),create.graph(B, list("Spearman")))
}
