% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNT-package.R
\docType{package}
\name{DNT-package}
\alias{DNT}
\alias{DNT-package}
\title{DNT: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
