% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{edge.inv}
\alias{edge.inv}
\title{Edge without direction}
\usage{
edge.inv(A, B)
}
\arguments{
\item{A}{a matrix.}

\item{B}{a matrix.}
}
\value{
Return
}
\description{
A
}
\details{
One of the methods to compare two networks on their edge-specific differneces.
}
\examples{
edge.inv(A,B)
}
