% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{eigen.inv}
\alias{eigen.inv}
\title{Calculates differences in eigenvector centrality for each network node}
\usage{
eigen.inv(X, Y)
}
\arguments{
\item{X, Y}{adjacency matrices of the same dimension}
}
\value{
a vector of length \eqn{N} (number of nodes), containing the differences in eigenvector centrality between the two networks for each node
}
\description{
This function calculates the differences in eigenvector centrality for each node between two networks (adjacency matrices of the same dimension).
}
\details{
This function calculates the differences in eigenvector centrality for each node between two networks (adjacency matrices of the same dimension). Differences in eigenvector centrality are one of the node-specific network difference characteristics to compare two networks.
}
\examples{
X<-create.adjacency.matrix(ExDataA,methodlist=list("Spearman"))
Y<-create.adjacency.matrix(ExDataB,methodlist=list("Spearman"))
eigen.inv(X,Y)
}
