% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExDataA-data.R
\docType{data}
\name{ExDataA}
\alias{ExDataA}
\title{Example data from which a network is constructed}
\format{
a matrix consisting of 130 rows (corresponding to samples) and 10 columns (corresponding to the network nodes)
}
\usage{
data(ExDataA)
}
\description{
example data set as a basis for constructing a network consisting of 10 nodes
}
\keyword{datasets}
