% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{frobenius.metric}
\alias{frobenius.metric}
\title{Calculates the Frobenius metric}
\usage{
frobenius.metric(A, B)
}
\arguments{
\item{A, B}{adjacency matrices of the same dimension}
}
\value{
the Frobenius metric between the two adjacency matrices \code{A} and \code{B}
}
\description{
This function calculates the Frobenius metric between two adjacency matrices of the same dimension.
}
\details{
This function calculates the Frobenius metric between two adjacency matrices of the same dimension. The Frobenius metric is an overall characteristic that can be employed to compare two networks.
}
\examples{
A<-create.adjacency.matrix(ExDataA,methodlist=list("Spearman"))
B<-create.adjacency.matrix(ExDataB,methodlist=list("Spearman"))
frobenius.metric(A,B)
}
