% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{global.str}
\alias{global.str}
\title{Global strength}
\usage{
global.str(A, B)
}
\arguments{
\item{A, B}{are matrices.}
}
\value{
The global strength of the two adjacency matrices \code{A} and \code{B}.
}
\description{
The method calculates the Global strength.
}
\details{
One of the methods to compare two networks on their overall network characteristics.
}
\examples{
global.str(matrix(3,2), matrix(2,2))

}
