% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R
\name{import.table}
\alias{import.table}
\title{Import table}
\usage{
import.table(directory)
}
\arguments{
\item{directory}{The directory to the file that should be loaded into R as a tibble.}
}
\value{
The data loaded out of the data file.
}
\description{
The function loads a table out of an extern file into R.
}
\details{
The function will only keep the columns with integers or doubles and the function can work with RData, text, excel, csv files.
}
