% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{jaccard.dist}
\alias{jaccard.dist}
\title{Calculates the Jaccard distance}
\usage{
jaccard.dist(A, B)
}
\arguments{
\item{A, B}{adjacency matrices of the same dimension}
}
\value{
the Jaccard distance between the two adjacency matrices \code{A} and \code{B}
}
\description{
This function calculates the Jaccard distance between two adjacency matrices of the same dimension.
}
\details{
This function calculates the Jaccard distance between two adjacency matrices of the same dimension. The Jaccard distance is an overall characteristic that can be employed to compare two networks.
}
\examples{
A<-create.adjacency.matrix(ExDataA,methodlist=list("Spearman"))
B<-create.adjacency.matrix(ExDataB,methodlist=list("Spearman"))
jaccard.dist(A,B)
}
