% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R
\name{layout.func}
\alias{layout.func}
\title{Creates layout function}
\usage{
layout.func(layout)
}
\arguments{
\item{layout}{the layout as a string}
}
\value{
the corresponding layout function
}
\description{
This function transforms a layout string from the igraph package into a corresponding layout function.
}
\details{
This function transforms a layout string from the igraph package into a corresponding layout function. If the string does not represent an existing layout from the igraph package, the layout function is automatically set to layout.auto.
}
