% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{number.differences}
\alias{number.differences}
\title{Calculates the differences with respect to the number of edges, clusters and isolated nodes between two networks}
\usage{
number.differences(A, B)
}
\arguments{
\item{A, B}{output results when applying the function \code{create.graph} to two data tables from which the two networks are constructed}
}
\value{
a vector of length 6 containing the graph difference in the number of edges, the graph difference in the number of clusters, the graph difference in the number of isolated nodes, the MST difference in the number of edges, the MST difference in the number of clusters and the MST difference in the number of isolated nodes
}
\description{
This function calculates the differences with respect to the number of edges, clusters (obtained by the Girvan-Newman algorithm) and isolated nodes (i.e. nodes without any edge) between two networks, for both \code{igraph} graphs and the corresponding minimum spanning trees (MSTs).
}
\details{
This function calculates the differences with respect to the number of edges, clusters (obtained by the Girvan-Newman algorithm) and isolated nodes (i.e. nodes without any edge) between two networks, for both \command{igraph} graphs and the corresponding minimum spanning trees (MSTs). It builds on the \command{create.graph} function, which creates \command{igraph} graph and corresponding MST results from adjacency matrices that are produced out of input tables, using an estimation method specified by the user; see the documentation of \command{create.graph} for further information. Differences in numbers of edges, clusters and isolated nodes are overall characteristics that can be employed to compare two networks.
}
\examples{
A<-create.graph(ExDataA,methodlist=list("Spearman"))
B<-create.graph(ExDataB,methodlist=list("Spearman"))
number.differences(A,B)
}
