% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{perm.test.nw}
\alias{perm.test.nw}
\title{Permutation test network}
\usage{
perm.test.nw(
  A,
  B,
  permnum,
  methodlist,
  thresh = 0.05,
  score.funct,
  paired = FALSE
)
}
\arguments{
\item{A}{a matrix.}

\item{B}{a matrix.}

\item{permnum}{the number of permutations.}

\item{methodlist}{the methods used to create the adjacency matrices.}

\item{score.funct}{the function used to compare the adjacency matrices A and B.}
}
\value{
Return
}
\description{
A
}
\details{
Details
}
\examples{
perm.test.nw(A = x1, B = x2, permnum = 10, methodlist = list("DistCorr"), thresh = 0.05, score.funct = frobenius.metric)
}
