% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{spec.dist}
\alias{spec.dist}
\title{Calculates the spectral distance}
\usage{
spec.dist(A, B)
}
\arguments{
\item{A, B}{adjacency matrices of the same dimension}
}
\value{
the spectral distance between the two adjacency matrices \code{A} and \code{B}
}
\description{
This function calculates the spectral distance between two adjacency matrices of the same dimension.
}
\details{
This function calculates the spectral distance between two adjacency matrices of the same dimension. The spectral distance is an overall characteristic that can be employed to compare two networks.
}
\examples{
A<-create.adjacency.matrix(ExDataA,methodlist=list("Spearman"))
B<-create.adjacency.matrix(ExDataB,methodlist=list("Spearman"))
spec.dist(A,B)
}
