% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{betweenness.inv}
\alias{betweenness.inv}
\title{Betweenness}
\usage{
betweenness.inv(X, Y)
}
\value{
Return
}
\description{
A
}
\details{
One of the methods to compare two networks on their node-specific differences.
}
\examples{
betweenness.inv(X,Y)

}
