% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{create.Igraphclustering}
\alias{create.Igraphclustering}
\title{Create cluster of an Igraph}
\usage{
create.Igraphclustering(A, methodlist, thresh = 0.05)
}
\arguments{
\item{A}{a matrix.}

\item{methodlist}{the methods used to create the adjacency matrix.}

\item{thresh}{the threshold under which the values of the correlations are relevant.}
}
\value{
The method returns a list with the Igraph from A and its clusters.
}
\description{
The method creates an Igraph out of the matrix A and the Igraphs clusters.
}
\details{
First the method creates the adjacency matrix out of the matrix A and after that the adjacency matrix can be
used to create the Igraph and its clusters.
}
