% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{degree.inv}
\alias{degree.inv}
\title{Calculates differences in degree for each network node}
\usage{
degree.inv(X, Y)
}
\arguments{
\item{X, Y}{adjacency matrices of the same dimension}
}
\value{
a vector of length \eqn{N} (number of nodes), containing the differences in degree between the two networks for each node
}
\description{
This function calculates the differences in degree for each node between two networks (adjacency matrices of the same dimension).
}
\details{
This function calculates the differences in degree for each node between two networks (adjacency matrices of the same dimension). Differences in degree are one of the node-specific network difference characteristics to compare two networks.
}
\examples{
X<-create.adjacency.matrix(ExDataA,methodlist=list("Spearman"))
Y<-create.adjacency.matrix(ExDataB,methodlist=list("Spearman"))
degree.inv(X,Y)
}
