% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNT-package.R
\docType{package}
\name{DNT-package}
\alias{DNT}
\alias{DNT-package}
\title{DNT: Differential network tests}
\description{
The DNT package provides a frame for testing for differences between two statistical networks based on permutation tests. Various options for network estimation (i.e. specifying the adjacency matrix) and network comparison (using several overall, node-specifc or edge-specific network difference chacteristics) are implemented. Moreover, tools for graphical illustrations and comparisons are offered.
}
\keyword{internal}
