% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNT-package.R
\docType{package}
\name{DNT-package}
\alias{DNT}
\alias{DNT-package}
\title{DNT: Differential network testing}
\description{
The package provides functions to test for differences between two statistical networks based on permutation tests. There are various options for network estimation and network comparison (using various types of overall, node-specifc or edge-specific network difference chacteristics). Moreover, tools for graphical illustrations and comparisons are offered.
}
\keyword{internal}
