% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{edge.inv.direc}
\alias{edge.inv.direc}
\title{Edge with direction}
\usage{
edge.inv.direc(A, B)
}
\arguments{
\item{A, B}{are matrices.}
}
\value{
Return
}
\description{
A
}
\details{
One of the methods to compare two networks on their edge-specific differneces.
}
\examples{
edge.inv.direc(A,B)
}
