% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{frobenius.metric}
\alias{frobenius.metric}
\title{Frobenius metric}
\usage{
frobenius.metric(A, B)
}
\arguments{
\item{A, B}{are matrices.}
}
\value{
The frobenius metric of the two adjacency matrices \code{A} and \code{B}.
}
\description{
The method calculates the Frobenius metric.
}
\details{
One of the methods to compare two networks on their overall network characteristics.
}
\examples{
frobenius.metric(matrix(3,2), matrix(2,2))
}
