% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{global.str}
\alias{global.str}
\title{Calculates the difference with respect to global strength}
\usage{
global.str(A, B)
}
\arguments{
\item{A, B}{adjacency matrices of the same dimension}
}
\value{
the difference with respect to the global strength between the two adjacency matrices \code{A} and \code{B}.
}
\description{
This function calculates the difference with respect to the global strength between two adjacency matrices of the same dimension.
}
\details{
This function calculates the difference with respect to the global strength between two adjacency matrices of the same dimension. The global strength is an overall characteristic that can be employed to compare two networks.
}
\examples{
A<-create.adjacency.matrix(ExDataA,methodlist=list("Spearman"))
B<-create.adjacency.matrix(ExDataB,methodlist=list("Spearman"))
global.str(A,B)
}
