% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{graph.plot}
\alias{graph.plot}
\title{Plots of the graphs}
\usage{
graph.plot(
  A,
  B,
  methodlist,
  thresh = 0.05,
  networkA = TRUE,
  networkB = TRUE,
  networkAtitle = "Network A",
  networkBtitle = "Network B",
  cluster = TRUE,
  negcol = "red",
  poscol = "blue",
  multiplier = 4,
  curved = TRUE,
  layout = igraph::layout.fruchterman.reingold,
  vSize = 16,
  tSize = 0.8
)
}
\arguments{
\item{A}{a matrix.}

\item{B}{a matrix.}

\item{methodlist}{the methods used to create the adjacency matrices.}

\item{thresh}{the threshold under which the values of the correlations are relevant.}

\item{networkA}{Boolean value that specifies if the Survivors Admission shall be plotted.}

\item{networkB}{Boolean value that specifies if the Survivors Event shall be plotted.}

\item{networkAtitle}{The title of the graph of the network of A.}

\item{networkBtitle}{The title of the graph of the network of B.}

\item{cluster}{is the boolean value that states if the graphs shall be divided in their clusters or not.}

\item{negcol}{the color of the edges with a negative correlation value.}

\item{poscol}{the color of the edges with a positive correlation value.}

\item{multiplier}{the number multiplied with the edge weight to get a bigger differnece between the thick and thin lines.}

\item{layout}{the layout of the networks.}

\item{vSize}{the size of the vertices.}

\item{tSize}{the size of the names inside the vertices.}
}
\value{

}
\description{
This method plots two networks to compare their differences.
}
\details{
The graphs are plotted with the same layout and other plot characteristics so that the two graphs are easy to
compare. The left network is the Survivors Admission and the right one the Survivor Event.
}
