% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R
\name{import.table}
\alias{import.table}
\title{Imports an external table into R}
\usage{
import.table(directory)
}
\arguments{
\item{directory}{the directory to the external table that should be imported into R}
}
\value{
the imported data table
}
\description{
This function imports an external table into R.
}
\details{
This function imports an external table into R. It causes that only columns with integers or doubles are kept and works with .RData, .txt, .xls and .csv files as input.
}
