% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R, R/app.R
\name{layout.func}
\alias{layout.func}
\title{Layout function}
\usage{
layout.func(layout)

layout.func(layout)
}
\arguments{
\item{layout}{The layout as a String.}
}
\value{
The suitable layout function.

The suitable layout function.
}
\description{
Layout function turns the layout-String into the suitable function.

Layout function turns the layout-String into the suitable function.
}
\details{
If the String is not a existing layout the layout function is automatically layout.auto.

If the String is not a existing layout the layout function is automatically layout.auto.
}
\examples{
layout.func("layout.fruchterman.reingold")
layout.func("layout.fruchterman.reingold")
}
