% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{maximum.metric}
\alias{maximum.metric}
\title{Maximum metric}
\usage{
maximum.metric(A, B)
}
\arguments{
\item{A, B}{matices.}
}
\value{
The Maximum metric of the two adjacency matrices \code{A} and \code{B}.
}
\description{
The method calculates the Maximum metric.
}
\details{
One of the methods to compare two networks on their overall network characteristics.
}
\examples{
maximum.metric(matrix(3,2), matrix(2,2))
}
