% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceCodePTNWNeu.R
\name{spec.dist}
\alias{spec.dist}
\title{Spectral distance}
\usage{
spec.dist(A, B)
}
\arguments{
\item{A, B}{are matrices.}
}
\value{
The spectral distance of the two adjacency matrices \code{A} and \code{B}.
}
\description{
The method calculates the Spectral distance.
}
\details{
One of the methods to compare two networks on their overall network characteristics.
}
\examples{
spec.dist(matrix(3,2), matrix(2,2))

}
